move clean_identifier to utils
