+++ garden_ai/app/pipeline.py
@@ -1,7 +1,5 @@
 import logging
-import re
 from datetime import datetime
-from keyword import iskeyword
 from pathlib import Path
 from typing import List, Optional
 
@@ -17,41 +15,11 @@ from garden_ai.utils.filesystem import (
     load_pipeline_from_python_file,
     PipelineLoadException,
 )
+from garden_ai.utils.misc import clean_identifier
 
 logger = logging.getLogger()
 
 pipeline_app = typer.Typer(name="pipeline", no_args_is_help=True)
-
-
-def clean_identifier(name: str) -> str:
-    """Clean the name provided for use as a pipeline's python identifier."""
-    orig = name
-    # Remove invalid characters, replacing with _
-    name = re.sub("[^0-9a-zA-Z_]", "_", name)
-
-    # Remove leading characters until we find a letter
-    name = re.sub("^[^a-zA-Z]+", "", name)
-
-    # Remove doubled/trailing underscores
-    name = re.sub("__+", "_", name).rstrip("_")
-
-    if not name:
-        # name consisted only of invalid characters
-        raise typer.BadParameter(
-            "Invalid shortname. This argument should contain a valid python identifier"
-            "(i.e. something usable as a variable name)."
-        )
-
-    # truncate to sane length, though not strictly necessary
-    name = name[:50]
-
-    if iskeyword(name):
-        name += "_"
-
-    if name != orig:
-        print(f'Generated valid shortname "{name}" from "{orig}".')
-
-    return name
 
 
 def parse_full_name(name: str) -> str:
