+++ garden_ai/client.py
@@ -134,8 +134,7 @@ class GardenClient:
                 AuthClient.scopes.openid,
                 AuthClient.scopes.email,
                 GroupsClient.scopes.view_my_groups_and_memberships,
-                SearchClient.scopes.ingest,
-                SearchClient.scopes.search,
+                SearchClient.scopes.all,
                 GardenClient.scopes.action_all,
                 Client.FUNCX_SCOPE,
             ],
